// src/API.ts
export type Admin = {
  __typename: "Admin",
  address?: string | null,
  assignedParkingIds?: Array< string | null > | null,
  buildingCode: string,
  buildingName?: string | null,
  buildingNo?: string | null,
  cognitoUsername?: string | null,
  createdAt: string,
  email: string,
  id: string,
  managerCode: string,
  managerName: string,
  phoneNo?: string | null,
  updatedAt: string,
};
// ...
export enum ParkingSpotSpotType {
  ELECTRIC = "ELECTRIC",
  GUEST = "GUEST",
  REGULAR = "REGULAR",
}
// ...
export type ModelAdminConditionInput = {
  address?: ModelStringInput | null,
  and?: Array< ModelAdminConditionInput | null > | null,
  assignedParkingIds?: ModelStringInput | null,
  buildingCode?: ModelStringInput | null,
  buildingName?: ModelStringInput | null,
  buildingNo?: ModelStringInput | null,
  cognitoUsername?: ModelStringInput | null,
  createdAt?: ModelStringInput | null,
  email?: ModelStringInput | null,
  managerCode?: ModelStringInput | null,
  managerName?: ModelStringInput | null,
  not?: ModelAdminConditionInput | null,
  or?: Array< ModelAdminConditionInput | null > | null,
  phoneNo?: ModelStringInput | null,
  updatedAt?: ModelStringInput | null,
};

export type CreateAdminInput = {
  address?: string | null,
  assignedParkingIds?: Array< string | null > | null,
  buildingCode: string,
  buildingName?: string | null,
  buildingNo?: string | null,
  cognitoUsername?: string | null,
  email: string,
  id?: string | null,
  managerCode: string,
  managerName: string,
  phoneNo?: string | null,
};

export type CreateAdminWithCognitoReturnType = {
  __typename: "CreateAdminWithCognitoReturnType",
  adminId?: string | null,
  cognitoUsername?: string | null,
  message: string,
  success: boolean,
  temporaryPassword?: string | null,
};
// ...
export type DeleteAdminInput = {
  id: string,
};
// ...
export type CreateAdminMutationVariables = {
  condition?: ModelAdminConditionInput | null,
  input: CreateAdminInput,
};

export type CreateAdminMutation = {
  createAdmin?:  {
    __typename: "Admin",
    address?: string | null,
    assignedParkingIds?: Array< string | null > | null,
    buildingCode: string,
    buildingName?: string | null,
    buildingNo?: string | null,
    cognitoUsername?: string | null,
    createdAt: string,
    email: string,
    id: string,
    managerCode: string,
    managerName: string,
    phoneNo?: string | null,
    updatedAt: string,
  } | null,
};

export type CreateAdminWithCognitoMutationVariables = {
  address?: string | null,
  buildingCode: string,
  buildingName?: string | null,
  buildingNo?: string | null,
  email: string,
  managerCode: string,
  managerName: string,
  phoneNo?: string | null,
};

export type CreateAdminWithCognitoMutation = {
  createAdminWithCognito?:  {
    __typename: "CreateAdminWithCognitoReturnType",
    adminId?: string | null,
    cognitoUsername?: string | null,
    message: string,
    success: boolean,
    temporaryPassword?: string | null,
  } | null,
};

export type CreateBuildingMutationVariables = {
  condition?: ModelBuildingConditionInput | null,
  input: CreateBuildingInput,
};

export type CreateBuildingMutation = {
  createBuilding?:  {
    __typename: "Building",
    address: string,
    buildingCode: string,
    buildingName: string,
    buildingNo: string,
    createdAt: string,
    id: string,
    location?: string | null,
    updatedAt: string,
  } | null,
};

export type CreateParkingMutationVariables = {
  condition?: ModelParkingConditionInput | null,
  input: CreateParkingInput,
};

export type CreateParkingMutation = {
  createParking?:  {
    __typename: "Parking",
    buildingCode: string,
    buildingName?: string | null,
    buildingNo?: string | null,
    createdAt: string,
    description?: string | null,
    id: string,
    parkingLots?: number | null,
    parkingName?: string | null,
    parkingNo: string,
    updatedAt: string,
  } | null,
};

export type CreateParkingSpotMutationVariables = {
  condition?: ModelParkingSpotConditionInput | null,
  input: CreateParkingSpotInput,
};

export type CreateParkingSpotMutation = {
  createParkingSpot?:  {
    __typename: "ParkingSpot",
    buildingCode: string,
    createdAt: string,
    description?: string | null,
    floor?: string | null,
    id: string,
    isAvailable?: boolean | null,
    parkingId: string,
    section?: string | null,
    spotNumber: string,
    spotType?: ParkingSpotSpotType | null,
    updatedAt: string,
  } | null,
};

export type CreateReservingMutationVariables = {
  condition?: ModelReservingConditionInput | null,
  input: CreateReservingInput,
};

export type CreateReservingMutation = {
  createReserving?:  {
    __typename: "Reserving",
    accessNo: string,
    buildingCode: string,
    createdAt: string,
    email?: string | null,
    endTime: string,
    id: string,
    phone?: string | null,
    reservedBy?: string | null,
    spotId: string,
    startTime: string,
    status?: ReservingStatus | null,
    updatedAt: string,
    vehicleCode?: string | null,
  } | null,
};

export type CreateUnitInfoMutationVariables = {
  condition?: ModelUnitInfoConditionInput | null,
  input: CreateUnitInfoInput,
};

export type CreateUnitInfoMutation = {
  createUnitInfo?:  {
    __typename: "UnitInfo",
    accessNo: string,
    buildingCode: string,
    cognitoUsername?: string | null,
    createdAt: string,
    email?: string | null,
    firstName: string,
    id: string,
    lastName: string,
    phone?: string | null,
    sakenLastName?: string | null,
    sakenName?: string | null,
    unitNo: string,
    updatedAt: string,
    username: string,
  } | null,
};

export type DeleteAdminMutationVariables = {
  condition?: ModelAdminConditionInput | null,
  input: DeleteAdminInput,
};

export type DeleteAdminMutation = {
  deleteAdmin?:  {
    __typename: "Admin",
    address?: string | null,
    assignedParkingIds?: Array< string | null > | null,
    buildingCode: string,
    buildingName?: string | null,
    buildingNo?: string | null,
    cognitoUsername?: string | null,
    createdAt: string,
    email: string,
    id: string,
    managerCode: string,
    managerName: string,
    phoneNo?: string | null,
    updatedAt: string,
  } | null,
};

export type DeleteAdminFromCognitoMutationVariables = {
  username: string,
};

export type DeleteAdminFromCognitoMutation = {
  deleteAdminFromCognito?:  {
    __typename: "DeleteAdminFromCognitoReturnType",
    message: string,
    success: boolean,
  } | null,
};

export type DeleteBuildingMutationVariables = {
  condition?: ModelBuildingConditionInput | null,
  input: DeleteBuildingInput,
};

export type DeleteBuildingMutation = {
  deleteBuilding?:  {
    __typename: "Building",
    address: string,
    buildingCode: string,
    buildingName: string,
    buildingNo: string,
    createdAt: string,
    id: string,
    location?: string | null,
    updatedAt: string,
  } | null,
};

export type DeleteParkingMutationVariables = {
  condition?: ModelParkingConditionInput | null,
  input: DeleteParkingInput,
};

export type DeleteParkingMutation = {
  deleteParking?:  {
    __typename: "Parking",
    buildingCode: string,
    buildingName?: string | null,
    buildingNo?: string | null,
    createdAt: string,
    description?: string | null,
    id: string,
    parkingLots?: number | null,
    parkingName?: string | null,
    parkingNo: string,
    updatedAt: string,
  } | null,
};

export type DeleteParkingSpotMutationVariables = {
  condition?: ModelParkingSpotConditionInput | null,
  input: DeleteParkingSpotInput,
};

export type DeleteParkingSpotMutation = {
  deleteParkingSpot?:  {
    __typename: "ParkingSpot",
    buildingCode: string,
    createdAt: string,
    description?: string | null,
    floor?: string | null,
    id: string,
    isAvailable?: boolean | null,
    parkingId: string,
    section?: string | null,
    spotNumber: string,
    spotType?: ParkingSpotSpotType | null,
    updatedAt: string,
  } | null,
};

export type DeleteReservingMutationVariables = {
  condition?: ModelReservingConditionInput | null,
  input: DeleteReservingInput,
};

export type DeleteReservingMutation = {
  deleteReserving?:  {
    __typename: "Reserving",
    accessNo: string,
    buildingCode: string,
    createdAt: string,
    email?: string | null,
    endTime: string,
    id: string,
    phone?: string | null,
    reservedBy?: string | null,
    spotId: string,
    startTime: string,
    status?: ReservingStatus | null,
    updatedAt: string,
    vehicleCode?: string | null,
  } | null,
};

export type DeleteUnitInfoMutationVariables = {
  condition?: ModelUnitInfoConditionInput | null,
  input: DeleteUnitInfoInput,
};

export type DeleteUnitInfoMutation = {
  deleteUnitInfo?:  {
    __typename: "UnitInfo",
    accessNo: string,
    buildingCode: string,
    cognitoUsername?: string | null,
    createdAt: string,
    email?: string | null,
    firstName: string,
    id: string,
    lastName: string,
    phone?: string | null,
    sakenLastName?: string | null,
    sakenName?: string | null,
    unitNo: string,
    updatedAt: string,
    username: string,
  } | null,
};

export type SuspendAdminInCognitoMutationVariables = {
  suspend: boolean,
  username: string,
};

export type SuspendAdminInCognitoMutation = {
  suspendAdminInCognito?:  {
    __typename: "SuspendAdminInCognitoReturnType",
    message: string,
    success: boolean,
  } | null,
};

export type UpdateAdminMutationVariables = {
  condition?: ModelAdminConditionInput | null,
  input: UpdateAdminInput,
};

export type UpdateAdminMutation = {
  updateAdmin?:  {
    __typename: "Admin",
    address?: string | null,
    assignedParkingIds?: Array< string | null > | null,
    buildingCode: string,
    buildingName?: string | null,
    buildingNo?: string | null,
    cognitoUsername?: string | null,
    createdAt: string,
    email: string,
    id: string,
    managerCode: string,
    managerName: string,
    phoneNo?: string | null,
    updatedAt: string,
  } | null,
};

export type UpdateBuildingMutationVariables = {
  condition?: ModelBuildingConditionInput | null,
  input: UpdateBuildingInput,
};

export type UpdateBuildingMutation = {
  updateBuilding?:  {
    __typename: "Building",
    address: string,
    buildingCode: string,
    buildingName: string,
    buildingNo: string,
    createdAt: string,
    id: string,
    location?: string | null,
    updatedAt: string,
  } | null,
};

export type UpdateParkingMutationVariables = {
  condition?: ModelParkingConditionInput | null,
  input: UpdateParkingInput,
};

export type UpdateParkingMutation = {
  updateParking?:  {
    __typename: "Parking",
    buildingCode: string,
    buildingName?: string | null,
    buildingNo?: string | null,
    createdAt: string,
    description?: string | null,
    id: string,
    parkingLots?: number | null,
    parkingName?: string | null,
    parkingNo: string,
    updatedAt: string,
  } | null,
};

export type UpdateParkingSpotMutationVariables = {
  condition?: ModelParkingSpotConditionInput | null,
  input: UpdateParkingSpotInput,
};

export type UpdateParkingSpotMutation = {
  updateParkingSpot?:  {
    __typename: "ParkingSpot",
    buildingCode: string,
    createdAt: string,
    description?: string | null,
    floor?: string | null,
    id: string,
    isAvailable?: boolean | null,
    parkingId: string,
    section?: string | null,
    spotNumber: string,
    spotType?: ParkingSpotSpotType | null,
    updatedAt: string,
  } | null,
};

export type UpdateReservingMutationVariables = {
  condition?: ModelReservingConditionInput | null,
  input: UpdateReservingInput,
};

export type UpdateReservingMutation = {
  updateReserving?:  {
    __typename: "Reserving",
    accessNo: string,
    buildingCode: string,
    createdAt: string,
    email?: string | null,
    endTime: string,
    id: string,
    phone?: string | null,
    reservedBy?: string | null,
    spotId: string,
    startTime: string,
    status?: ReservingStatus | null,
    updatedAt: string,
    vehicleCode?: string | null,
  } | null,
};